test: cover course player quiz history UI
